fixed review filtering and notifications
